Update the project structure
As the project is becoming bigger and bigger
This new structure will help us to organize our work
app.rb has bee split in in several file in the folder controllers
