+++ Rakefile
@@ -1,5 +1,7 @@
 require 'rake/testtask'
-require './app'
+Dir.glob('./{config,models,services,controllers}/init.rb').each do |file|
+  require file
+end
 
 task :default => [:spec]
