+++ config/init.rb
@@ -1,0 +1,1 @@
+require_relative 'environments'
