+++ controllers/init.rb
@@ -1,0 +1,3 @@
+require_relative 'base'
+require_relative 'campaign_controller'
+require_relative 'tracker_controller'
