+++ models/init.rb
@@ -1,4 +1,4 @@
-require_relative './lib/init'
-require_relative './tracker'
-require_relative './campaign'
-require_relative './visit'
+require_relative 'lib/init'
+require_relative 'tracker'
+require_relative 'campaign'
+require_relative 'visit'
