+++ specs/spec_helper.rb
@@ -2,7 +2,10 @@ ENV['RACK_ENV'] = 'test'
 
 require 'minitest/autorun'
 require 'rack/test'
-require_relative '../app'
+
+Dir.glob('./{config,models,services,controllers}/init.rb').each do |file|
+  require file
+end
 
 include Rack::Test::Methods
